Eliminate race condition when triggering once to send border
It was possible that two callback is executed concurrently for receiving
border info and both of them would send the border.
Until a single task is executed at a time with potential overlap
with another task, there is no need to use lock to check whether the
border is sent, if there is no `await` between checking and setting the
value since the execution of a task will not be interrupted until an
`await` expression.