Wait for iteration before setting again border info for the same direction
An iteration of the system shall be initiated by sending the border
of a cell matrix to its neighbors.
If a cell matrix has iterated, it will not know the status of its
neighbors before the iteration is re-initiated.
This feature synchronizes with neighbors to ensure they are at the same
iteration when their border info is delivered.